removed default navItems, made navItems for auth and unauth props
